requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=0.23.0=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.11.6=pypi_0
aisuite=0.1.8=pypi_0
annotated-types=0.5.0=pypi_0
anyio=3.7.1=pypi_0
appnope=0.1.3=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.4.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=24.2.0=pypi_0
auto-gptq=0.4.2=pypi_0
autocommand=2.2.2=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
backoff=2.2.1=pypi_0
backports-tarfile=1.2.0=pypi_0
bcrypt=4.2.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
biothings-client=0.4.1=pypi_0
bleach=6.0.0=pypi_0
blis=0.7.11=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2024.7.4=hbcca054_0
catalogue=2.0.10=pypi_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
cheroot=10.0.1=pypi_0
cherrypy=18.10.0=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.4.13=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
cmake=3.27.6=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.1.4=pypi_0
confection=0.1.3=pypi_0
contextvars=2.4=pypi_0
contourpy=1.2.1=pypi_0
cramjam=2.9.1=pypi_0
cryptography=43.0.3=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
dataclasses-json=0.6.0=pypi_0
datasets=2.14.5=pypi_0
debugpy=1.8.0=pypi_0
dill=0.3.7=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
en-core-web-sm=3.7.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=1.2.0=pypi_0
fastapi=0.103.1=pypi_0
fastjsonschema=2.18.0=pypi_0
fastparquet=2024.11.0=pypi_0
filelock=3.12.4=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.53.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2023.6.0=pypi_0
future=1.0.0=pypi_0
fuzzywuzzy=0.18.0=pypi_0
greenlet=2.0.2=pypi_0
httptools=0.6.0=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.17.2=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
immutables=0.21=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.1.0=pypi_0
ipykernel=6.25.2=pypi_0
ipython=8.15.0=pypi_0
jaraco-collections=5.1.0=pypi_0
jaraco-context=6.0.1=pypi_0
jaraco-functools=4.1.0=pypi_0
jaraco-text=4.0.0=pypi_0
jedi=0.19.0=pypi_0
jinja2=3.1.4=pypi_0
jiter=0.8.2=pypi_0
jmespath=1.0.1=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
junos-eznc=2.7.2=pypi_0
jupyter-client=8.3.1=pypi_0
jupyter-core=5.3.1=pypi_0
jupyter-events=0.7.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.7.3=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.6=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.25.0=pypi_0
kiwisolver=1.4.5=pypi_0
langchain=0.0.301=pypi_0
langcodes=3.3.0=pypi_0
langsmith=0.0.40=pypi_0
ld_impl_linux-64=2.40=hf3520f5_7
levenshtein=0.26.0=pypi_0
libffi=3.4.2=h7f98852_5
libgcc-ng=14.1.0=h77fa898_0
libgomp=14.1.0=h77fa898_0
libnsl=2.0.1=hd590300_0
libsqlite=3.46.0=hde9e2c9_0
libuuid=2.38.1=h0b41bf4_0
libzlib=1.3.1=h4ab18f5_1
lit=17.0.2=pypi_0
litellm=1.59.12=pypi_0
llama-cpp-python=0.2.7=pypi_0
looseversion=1.3.0=pypi_0
lxml=5.3.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.20.1=pypi_0
matplotlib=3.9.1=pypi_0
matplotlib-venn=1.1.1=pypi_0
mistune=3.0.1=pypi_0
monotonic=1.6=pypi_0
more-itertools=10.5.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.15=pypi_0
murmurhash=1.0.10=pypi_0
mygene=3.2.2=pypi_0
napalm=5.0.0=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.8.0=pypi_0
ncclient=0.6.15=pypi_0
ncurses=6.5=h59595ed_0
neo4j=5.12.0=pypi_0
nest-asyncio=1.5.8=pypi_0
netaddr=1.3.0=pypi_0
netmiko=4.4.0=pypi_0
netutils=1.10.0=pypi_0
networkx=3.1=pypi_0
notebook=7.0.4=pypi_0
ntc-templates=7.4.0=pypi_0
numexpr=2.8.6=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
onnxruntime=1.16.0=pypi_0
openai=1.61.0=pypi_0
openssl=3.3.1=h4ab18f5_1
overrides=7.4.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.0=pypi_0
paramiko=3.5.0=pypi_0
patsy=1.0.1=pypi_0
peft=0.5.0=pypi_0
pexpect=4.8.0=pypi_0
pickled=2020.5.0a1=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.1=pypi_0
pip=25.0=pypi_0
pipdeptree=2.25.0=pypi_0
platformdirs=3.10.0=pypi_0
portend=3.2.0=pypi_0
posthog=3.0.2=pypi_0
preshed=3.0.9=pypi_0
progressbar2=4.5.0=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
propcache=0.2.0=pypi_0
protobuf=4.24.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pulsar-client=3.3.0=pypi_0
pyarrow=13.0.0=pypi_0
pycparser=2.22=pypi_0
pycryptodomex=3.21.0=pypi_0
pydantic=2.4.0=pypi_0
pydantic-core=2.10.0=pypi_0
pyeapi=1.0.4=pypi_0
pygments=2.16.1=pypi_0
pynacl=1.5.0=pypi_0
pyopenssl=24.2.1=pypi_0
pyparsing=3.2.0=pypi_0
pypika=0.48.9=pypi_0
pyserial=3.5=pypi_0
python=3.10.9=he550d4f_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.0=pypi_0
python-gnupg=0.5.3=pypi_0
python-graphviz=0.21=pypi_0
python-levenshtein=0.26.0=pypi_0
python-utils=3.9.0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.1=pypi_0
rapidfuzz=3.10.0=pypi_0
readline=8.2=h8228510_1
referencing=0.30.2=pypi_0
regex=2023.8.8=pypi_0
requests=2.32.3=pypi_0
rouge=1.0.1=pypi_0
rpds-py=0.10.3=pypi_0
rpm-vercmp=0.1.2=pypi_0
sacremoses=0.0.53=pypi_0
safetensors=0.3.3=pypi_0
salt=3007.1=pypi_0
salt-sproxy=2023.8.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
scp=0.15.0=pypi_0
seaborn=0.13.2=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=70.1.1=pyhd8ed1ab_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
spacy=3.7.2=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlalchemy=2.0.21=pypi_0
srsly=2.4.8=pypi_0
stack-data=0.6.2=pypi_0
starlette=0.27.0=pypi_0
statsmodels=0.14.4=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tempora=5.7.0=pypi_0
terminado=0.17.1=pypi_0
textfsm=1.1.3=pypi_0
thinc=8.2.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.7.0=pypi_0
timelib=0.3.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.13.3=pypi_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.3.3=pypi_0
traitlets=5.10.0=pypi_0
transformers=4.33.2=pypi_0
transitions=0.9.2=pypi_0
triton=2.0.0=pypi_0
ttp=0.9.5=pypi_0
ttp-templates=0.3.7=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
uniprot-id-mapper=1.1.2=pypi_0
uniprotmapper=0.3.dev0=pypi_0
urllib3=2.2.3=pypi_0
uvicorn=0.23.2=pypi_0
uvloop=0.17.0=pypi_0
wasabi=1.1.2=pypi_0
watchfiles=0.20.0=pypi_0
wcwidth=0.2.6=pypi_0
weasel=0.3.3=pypi_0
websocket-client=1.6.3=pypi_0
websockets=11.0.3=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xxhash=3.4.1=pypi_0
xz=5.2.6=h166bdaf_0
yamlordereddictloader=0.4.2=pypi_0
yarl=1.17.2=pypi_0
zc-lockfile=3.0.post1=pypi_0
zipp=3.21.0=pypi_0
